Refactor combat logic into modular structure
