+++ src/app/game/combat/passives.js
@@ -1,0 +1,9 @@
+let passiveHandler = () => {};
+
+export function registerPassiveHandler(handler) {
+  passiveHandler = handler;
+}
+
+export function triggerAttackPassive(creature, controllerIndex) {
+  passiveHandler(creature, controllerIndex, 'onAttack');
+}
